Fixed warning on windows.
git-svn-id: https://svn.opendtect.org/od/trunk@27060 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Seis/seiscubeprov.cc
@@ -202,7 +202,7 @@ int SeisMSCProvider::estimatedNrTraces() const
     if ( estnrtrcs_ != -2 ) return estnrtrcs_;
     estnrtrcs_ = -1;
     if ( !rdr_.selData() )
-	return is2D() ? estnrtrcs_ : SI().sampling(false).hrg.totalNr();
+	return is2D() ? estnrtrcs_ : (int) SI().sampling(false).hrg.totalNr();
 
     estnrtrcs_ = rdr_.selData()->expectedNrTraces( is2D() );
     return estnrtrcs_;
